+++ test/entrypoint/event_channel.dart/throws.dart
@@ -4,7 +4,7 @@ import 'package:isolate_channel/isolate_channel.dart';
 
 void main(List<String> args, SendPort send) => throwsEntryPoint(send);
 
-void throwsEntryPoint(SendPort send) {
+void throwsEntryPoint(SendPort? send) {
   final connection = setupIsolate(send);
 
   final channel = IsolateEventChannel('test', connection);