+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 0.6.0
 
 - Removes `onConnect` callbacks in favor of `onSendPortReady` in `setupIsolate` to support headless restarts
-- The `setupIsolate` function now accepts a nullable `send` parameter to support headless restarts
+- The `IsolateEntryPoint` and `setupIsolate` functions now accept a nullable `send` parameter to support headless restarts
+- The `connectToIsolate` function now supports `onExit` and `onError` callback setup
 
 ## 0.5.0